+++ src/main/java/com/dasgupta/careercompass/jobApplication/JobApplicationRepository.java
@@ -2,5 +2,5 @@ package com.dasgupta.careercompass.jobApplication;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobApplicationRepository extends JpaRepository<JobApplication, Long> {
+public interface JobApplicationRepository extends JpaRepository<JobApplication, Integer> {
 }
